requirements.txt
# Function dependencies, for example:
# package>=version
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cloudevents==1.2.0
deprecation==2.1.0
Flask==2.0.2
functions-framework==2.3.0
google-api-core==2.2.0
google-auth==2.3.2
google-cloud==0.34.0
google-cloud-aiplatform==1.6.1
google-cloud-bigquery==2.29.0
google-cloud-core==2.1.0
google-cloud-language==2.3.2
google-cloud-storage==1.42.3
google-cloud-vision==2.6.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
opencv-contrib-python==4.5.4.58
packaging==21.0
proto-plus==1.19.7
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.1
python-dateutil==2.8.2
requests==2.26.0
rsa==4.7.2
scipy==1.7.1
six==1.16.0
urllib3==1.26.7
watchdog==1.0.2
Werkzeug==2.0.2